Consolidate torchrun acknowledgement reads
